cw1_cassandra/requirements.txt
click
cassandra-driver
voluptuous
